refactor: split awake.py into bridge_state.py + command_handlers.py
Extract shared config (bridge_state.py) and command handlers
(command_handlers.py) from awake.py to reduce module size from
962 to ~520 lines. Uses callback injection to avoid circular imports.

Supersedes PR #141 (split-awake-v2) which had conflicts with
upstream/main after PR #123 (/restart command) was merged.
