+++ koan/tests/test_restart.py
@@ -115,69 +115,69 @@ class TestRestartExitCode:
 class TestHandleRestartCommand:
     """Test that /restart triggers restart flow, not resume."""
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_restart_creates_signal_file(self, mock_send, tmp_path):
-        with patch("app.awake.KOAN_ROOT", tmp_path):
-            from app.awake import handle_restart
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+            from app.command_handlers import handle_restart
             handle_restart()
         assert (tmp_path / RESTART_FILE).exists()
         mock_send.assert_called_once()
         assert "Restart" in mock_send.call_args[0][0]
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_restart_clears_pause_state(self, mock_send, tmp_path):
         pause_file = tmp_path / ".koan-pause"
         reason_file = tmp_path / ".koan-pause-reason"
         pause_file.write_text("PAUSE")
         reason_file.write_text("manual")
 
-        with patch("app.awake.KOAN_ROOT", tmp_path):
-            from app.awake import handle_restart
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+            from app.command_handlers import handle_restart
             handle_restart()
 
         assert not pause_file.exists()
         assert not reason_file.exists()
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_restart_dedup_skips_when_file_exists(self, mock_send, tmp_path):
         """Second /restart call is a no-op when file already exists (dedup)."""
         (tmp_path / RESTART_FILE).write_text("already pending")
-        with patch("app.awake.KOAN_ROOT", tmp_path):
-            from app.awake import handle_restart
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+            from app.command_handlers import handle_restart
             handle_restart()
         # Should not send telegram (dedup — already pending)
         mock_send.assert_not_called()
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_restart_works_after_stale_file_cleared(self, mock_send, tmp_path):
         """After main() clears the stale file, new /restart is honored."""
         # Simulate: stale file was cleared by main() after first poll
         # (no file present)
-        with patch("app.awake.KOAN_ROOT", tmp_path):
-            from app.awake import handle_restart
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+            from app.command_handlers import handle_restart
             handle_restart()
         assert (tmp_path / RESTART_FILE).exists()
         mock_send.assert_called_once()
         assert "Restart" in mock_send.call_args[0][0]
 
-    @patch("app.awake.handle_restart")
+    @patch("app.command_handlers.handle_restart")
     def test_command_routes_restart_to_handler(self, mock_restart):
-        from app.awake import handle_command
+        from app.command_handlers import handle_command
         handle_command("/restart")
         mock_restart.assert_called_once()
 
-    @patch("app.awake.handle_resume")
+    @patch("app.command_handlers.handle_resume")
     def test_restart_does_not_call_resume(self, mock_resume, tmp_path):
-        with patch("app.awake.send_telegram"):
-            with patch("app.awake.KOAN_ROOT", tmp_path):
-                from app.awake import handle_command
+        with patch("app.command_handlers.send_telegram"):
+            with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+                from app.command_handlers import handle_command
                 handle_command("/restart")
         mock_resume.assert_not_called()
 
-    @patch("app.awake.handle_resume")
+    @patch("app.command_handlers.handle_resume")
     def test_resume_aliases_still_work(self, mock_resume):
         """Verify /work, /awake, /start still call resume, not restart."""
-        from app.awake import handle_command
+        from app.command_handlers import handle_command
         for cmd in ["/resume", "/work", "/awake", "/start"]:
             mock_resume.reset_mock()
             handle_command(cmd)
@@ -187,7 +187,7 @@ class TestHandleRestartCommand:
 class TestRestartLoopPrevention:
     """Verify the dedup mechanism that prevents infinite restart loops."""
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_stale_file_does_not_trigger_restart(self, mock_send, tmp_path):
         """A .koan-restart file from a previous incarnation is ignored."""
         # Create file in the past (previous incarnation)
@@ -200,7 +200,7 @@ class TestRestartLoopPrevention:
         startup_time = time.time()
         assert check_restart(tmp_path, since=startup_time) is False
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_fresh_file_triggers_restart(self, mock_send, tmp_path):
         """A new .koan-restart file (after startup) triggers restart."""
         startup_time = time.time() - 10
@@ -209,7 +209,7 @@ class TestRestartLoopPrevention:
         request_restart(tmp_path)
         assert check_restart(tmp_path, since=startup_time) is True
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_redelivered_restart_is_deduplicated(self, mock_send, tmp_path):
         """Simulates the restart loop scenario:
         1. /restart creates file
@@ -220,15 +220,15 @@ class TestRestartLoopPrevention:
         6. Future /restart works normally
         """
         # Step 1: First /restart
-        with patch("app.awake.KOAN_ROOT", tmp_path):
-            from app.awake import handle_restart
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
+            from app.command_handlers import handle_restart
             handle_restart()
         assert (tmp_path / RESTART_FILE).exists()
         assert mock_send.call_count == 1
 
         # Step 3-4: Re-delivered /restart — file still exists → dedup
         mock_send.reset_mock()
-        with patch("app.awake.KOAN_ROOT", tmp_path):
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
             handle_restart()
         mock_send.assert_not_called()
 
@@ -238,24 +238,24 @@ class TestRestartLoopPrevention:
 
         # Step 6: New /restart is now honored
         mock_send.reset_mock()
-        with patch("app.awake.KOAN_ROOT", tmp_path):
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path):
             handle_restart()
         assert (tmp_path / RESTART_FILE).exists()
         mock_send.assert_called_once()
 
 
 class TestHelpIncludesRestart:
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_help_lists_restart_command(self, mock_send):
-        from app.awake import _handle_help
+        from app.command_handlers import _handle_help
         _handle_help()
         help_text = mock_send.call_args[0][0]
         assert "/restart" in help_text
         assert "restart both" in help_text.lower()
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_help_resume_aliases_exclude_restart(self, mock_send):
-        from app.awake import _handle_help
+        from app.command_handlers import _handle_help
         _handle_help()
         help_text = mock_send.call_args[0][0]
         # Find the /resume line and check /restart is not listed as alias
